Refactor VerifyPhoneNumber to be PSR-16-compliant
Looking at the code, it struck me that it doesn't need to, and
shouldn't, be developed against a specific caching implementation (and
that the composer.json suggests update that was made in an earlier
commit was not fully helpful). So, this change refactors
VerifyPhoneNumber to be PSR-16 compliant by having it accept a
CacheInterface object for the cache, instead of a StorageInterface
object.