Move the query parameter filter functionality to a dedicated class
After looking at the new filter functionality, while simplistic, it
worked. However, it seemed to both be in the wrong place and not as
maintainable as it could be. So, I decided to start refactoring it using
laminas-inputfilter. This change is the first step in that direction.
I'm planning to continue refactoring it to better and more meaningfully
use laminas-inputfilter in future commits.