Add a custom laminas-filter to explode and filter string data
As the Field value of the Lookup V2 API is a comma-separated string, I
needed a simplistic way to filter out invalid options. So, I took
inspiration from laminas-validator's Explode validator and implemented a
simplistic FilterInstance class that explodes the provided value to the
filter and applies a filter on it before imploding the filtered data and
returning it.