Add customer laminas-inputfilter for filtering Lookup query parameters
This is, hopefully, the final approach that I'm taking to implementing
filtering of query parameters for the Lookup validator. Instead of
performing the filtering and validation of query parameters within the
validator itself, this custom InputFilter class is there to do that
externally. It will filter out any invalid input, where possible, and
then validate the remaining query parameters.