Change QueryParametersInputFilter input item names
This change changes the names of all of the input items that
QueryParametersInputFilter checks to match the query parameters that
Twilio's Lookup (V2) API accepts.

I don't know if there was a particular reason that I didn't do this
originally, but it doesn't make sense. There's no need to hide or
obfuscate the field names out of any sense of security, and I don't see
any other reason either.

So, this change updates them. Now, there doesn't need to be any mapping
functionality added, post-filter and validation.