Set and retrieve VerifyPhoneNumber query parameters
Initially the class only accepted query parameters at initialisation.
However, as I plan to integrate this package into
mezzio-twilio-integration, that will be too limiting.

So, this change adds the ability to set them post-initialisation, and to
retrieve them for validation and testing. That way, an instantiated
class can be retrieved from a DI container and still have query
parameters set, giving it more flexibility.