Improve the code documentation
I don't want to go over the top in this change, but I still want to
provide sufficient code documentation such that the code is as easy to
come up to speed with and maintain as possible. So, these comments have
been added to give a quick overview of the respective classes and
variables.