Add a TwilioRestClient and instantiating factory class
A Twilio HTTP Rest Client is likely to become ever more central to this
package. So it makes sense that it is both present and that it is as
easy to instantiate as possible. So, that's what this change does. In
addition, it introduces a ConfigProvider class to auto wire the class as
a service when used in Mezzio applications.