Fix bug in the VerifyPhoneNumberFactory
I didn't see this earlier, but there was a bug in the
VerifyPhoneNumberFactory where it attempted to retrieve a
TwilioRestClientFactory from the DI container instead of a Twilio HTTP
Rest Client. I can only guess that it got through because there wasn't a
test for the class up until now.

So, this change introduces a covering test class and corrects the bug in
VerifyPhoneNumberFactory.